requirements-test.txt
-r requirements.txt
mock
pytest
pytest-httpretty
